requirements.txt
Django==4.2.13
django_tinymce==4.0.0
python-dotenv==1.0.1
django-bootstrap-v5==1.0.11
psycopg2-binary==2.9.9